requirements.txt
asgiref==3.8.1
Django==5.2.3
sqlparse==0.5.3
# Add to requirements.txt
gunicorn==21.2.0  # or latest stable

python-dotenv==1.1.1
paynow==1.0.8
pgvector @ git+https://github.com/pgvector/pgvector-python.git@e211ba4029f204734f0c001fbb90f6a594d561ae
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
channels==4.2.2
channels_redis==4.2.1
requests==2.32.5
psycopg2-binary==2.9.10
stripe==12.5.0
sympy==1.14.0
syncedlyrics==1.0.1
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.4
Twisted==25.5.0
txaio==25.6.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
unstructured==0.18.15
unstructured-client==0.42.3
unstructured.pytesseract==0.3.15
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.23.2
vaderSentiment==3.3.2
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.5
xlrd==2.0.2
xxhash==3.6.0
yarl==1.20.1
yt-dlp==2025.6.30
ytmusicapi==1.10.3
zipp==3.23.0
zope.interface==7.2
zstandard==0.25.0

gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2